SYSTEM_BACKUP_2025-09-26_23-29-25/requirements.txt
# MEM_AGENT Python Dependencies

# Core Framework
streamlit>=1.28.0
fastapi>=0.104.0
uvicorn>=0.24.0

# Google Cloud Integration
google-cloud-secret-manager>=2.16.0
google-genai>=0.3.0
google-cloud-aiplatform>=1.38.0

# Database Integration
supabase>=2.0.0
psycopg2-binary>=2.9.0
redis>=5.0.0

# AI and ML
openai>=1.3.0
anthropic>=0.7.0
pandas>=2.1.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Web and API
requests>=2.31.0
httpx>=0.25.0
beautifulsoup4>=4.12.0
selenium>=4.15.0

# Data Processing
PyPDF2>=3.0.0
python-docx>=1.1.0
openpyxl>=3.1.0
Pillow>=10.0.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.5.0
loguru>=0.7.0
schedule>=1.2.0

# Development
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
